+++ config_reader.py
@@ -18,11 +18,10 @@ class Settings(BaseSettings):
   tts_replacements: Dict
   tts_credits: str
   tts_ffmpeg_path: str
-  tts_enable_so_vits_svc: bool
+  tts_enable_backends: List[Literal['say_macos', 'ttsx4', 'coqui_tts', 'so_vits_svc']]
   tts_list_system_voices: bool
   tts_so_vits_svc_4_0_code_path: str
   tts_so_vits_svc_4_1_code_path: str
-  tts_so_vits_svc_base_tts_provider: Literal["say_macos", "built_in"]
   tts_so_vits_svc_voices: List[Dict]
   tts_queue_size_per_user: int
   tts_host: str
@@ -93,17 +92,6 @@ class Settings(BaseSettings):
   def resolution_in_correct_ranges(cls, v):
     if v % 64 != 0 or v < 256 or v > 2048:
       raise ValueError('incorrect value')
-    return v
-  
-  @validator('tts_so_vits_svc_voices')
-  def base_voices_exist(cls, v, values):
-    if not values['tts_enable_so_vits_svc']:
-      return v
-    for item in v:
-      provider = item.get('provider', values['tts_so_vits_svc_base_tts_provider'])
-      if provider == 'built_in':
-        if item['base_voice'] not in values['tts_voices']:
-          raise ValueError(f'base tts voice ({item["base_voice"]}) does not exist')
     return v
   
   class Config: